Adjusted code examples formatting
