+++ docs/src/Demo/Demo.css
@@ -36,3 +36,4 @@ main {
     flex-basis: 100%;
   }
 }
+
