+++ docs/src/Demo/Examples/BasicUsage.jsx
@@ -1,11 +1,11 @@
 import React, { Component } from 'react';
-// import moment from 'moment';
+import moment from 'moment';
 import { Typography } from 'material-ui';
 import { TimePicker, DatePicker } from 'material-ui-pickers';
 
 export default class BasicUsage extends Component {
   state = {
-    // selectedDate: moment(),
+    selectedDate: moment(),
   }
 
   handleDateChange = (date) => {
