+++ docs/src/Demo/components/Code.jsx
@@ -1,0 +1,172 @@
+import React from 'react';
+import { withStyles } from 'material-ui/styles';
+
+const anchorLinkStyle = (theme, size) => ({
+  '& .anchor-link-style': {
+    opacity: 0,
+    // To prevent the link to get the focus.
+    display: 'none',
+  },
+  '&:hover .anchor-link-style': {
+    display: 'inline-block',
+    opacity: 1,
+    padding: `0 ${theme.spacing.unit}px`,
+    color: theme.palette.text.hint,
+    '&:hover': {
+      color: theme.palette.text.secondary,
+    },
+    '& svg': {
+      width: size,
+      fill: 'currentColor',
+    },
+  },
+});
+
+const styles = theme => ({
+  root: {
+    fontFamily: theme.typography.fontFamily,
+    fontSize: 16,
+    color: theme.palette.text.primary,
+    margin: 0,
+    backgroundColor: theme.palette.background.paper,
+
+    '& .anchor-link': {
+      marginTop: -theme.spacing.unit * 12, // Offset for the anchor.
+      position: 'absolute',
+    },
+    '& pre': {
+      borderRadius: 3,
+      overflow: 'auto',
+      margin: 0,
+      backgroundColor: theme.palette.background.paper,
+    },
+    '& code': {
+      display: 'inline-block',
+      lineHeight: 1.6,
+      fontFamily: 'Consolas, "Liberation Mono", Menlo, Courier, monospace',
+      padding: '3px 6px',
+      color: theme.palette.text.primary,
+      fontSize: 14,
+      backgroundColor: theme.palette.background.paper,
+    },
+    '& p code, & ul code, & pre code': {
+      fontSize: 14,
+      lineHeight: 1.6,
+    },
+    '& h1 code, & h2 code, & h3 code, & h4 code': {
+      fontSize: 'inherit',
+      lineHeight: 'inherit',
+    },
+    '& h1': {
+      ...theme.typography.display2,
+      color: theme.palette.text.secondary,
+      margin: '0.7em 0',
+      ...anchorLinkStyle(theme, 20),
+    },
+    '& h2': {
+      ...theme.typography.display1,
+      color: theme.palette.text.secondary,
+      margin: '1em 0 0.7em',
+      ...anchorLinkStyle(theme, 18),
+    },
+    '& h3': {
+      ...theme.typography.headline,
+      color: theme.palette.text.secondary,
+      margin: '1em 0 0.7em',
+      ...anchorLinkStyle(theme, 16),
+    },
+    '& h4': {
+      ...theme.typography.title,
+      color: theme.palette.text.secondary,
+      margin: '1em 0 0.7em',
+      ...anchorLinkStyle(theme, 14),
+    },
+    '& p, & ul, & ol': {
+      lineHeight: 1.6,
+    },
+    '& table': {
+      width: '100%',
+      display: 'block',
+      overflowX: 'auto',
+      borderCollapse: 'collapse',
+      borderSpacing: 0,
+      overflow: 'hidden',
+    },
+    '& thead': {
+      fontSize: 12,
+      fontWeight: theme.typography.fontWeightMedium,
+      color: theme.palette.text.secondary,
+    },
+    '& tbody': {
+      fontSize: 13,
+      lineHeight: 1.5,
+      color: theme.palette.text.primary,
+    },
+    '& td': {
+      borderBottom: `1px solid ${theme.palette.text.lightDivider}`,
+      padding: `${theme.spacing.unit}px ${theme.spacing.unit * 5}px ${theme.spacing.unit}px ${theme
+        .spacing.unit * 3}px`,
+      textAlign: 'left',
+    },
+    '& td:last-child': {
+      paddingRight: theme.spacing.unit * 3,
+    },
+    '& td compact': {
+      paddingRight: theme.spacing.unit * 3,
+    },
+    '& td code': {
+      fontSize: 13,
+      lineHeight: 1.6,
+    },
+    '& th': {
+      whiteSpace: 'pre',
+      borderBottom: `1px solid ${theme.palette.text.lightDivider}`,
+      padding: `0 ${theme.spacing.unit * 5}px 0 ${theme.spacing.unit * 3}px`,
+      textAlign: 'left',
+    },
+    '& th:last-child': {
+      paddingRight: theme.spacing.unit * 3,
+    },
+    '& tr': {
+      height: 48,
+    },
+    '& thead tr': {
+      height: 64,
+    },
+    '& strong': {
+      fontWeight: theme.typography.fontWeightMedium,
+    },
+    '& blockquote': {
+      borderLeft: `5px solid ${theme.palette.text.hint}`,
+      background: theme.palette.background.paper,
+      padding: `${theme.spacing.unit / 2}px ${theme.spacing.unit * 3}px`,
+      margin: `${theme.spacing.unit * 3}px 0`,
+    },
+    '& a, & a code': {
+      // Style taken from the Link component
+      color: theme.palette.secondary.A400,
+      textDecoration: 'none',
+      '&:hover': {
+        textDecoration: 'underline',
+      },
+    },
+  },
+});
+
+const Code = (props) => {
+  const {
+    classes, className, text, ...other
+  } = props;
+
+  return (
+    <div className={classes.root}>
+      <pre>
+        <code className="language-jsx">
+          {text}
+        </code>
+      </pre>
+    </div>
+  );
+};
+
+export default withStyles(styles)(Code);
