+++ docs/src/index.js
@@ -1,5 +1,4 @@
 /* eslint-disable no-use-before-define */
-
 import React from 'react';
 import ReactDOM from 'react-dom';
 import moment from 'moment';
