+++ docs/src/utils/prism.js
@@ -1,8 +1,4 @@
 import prism from 'prismjs';
-import 'prismjs/components/prism-markup';
-import 'prismjs/components/prism-javascript';
-import 'prismjs/components/prism-css';
-import 'prismjs/components/prism-diff';
 import 'prismjs/components/prism-jsx';
 
 /* eslint-disable import/no-webpack-loader-syntax  */
@@ -23,5 +19,3 @@ export function setPrismTheme(theme) {
     ? lightTheme
     : darkTheme;
 }
-
-export default prism;
